+++ !Source Code/Constants.h
@@ -1,0 +1,6 @@
+#ifndef CONSTANT_H_
+#define CONSTANT_H_
+
+#include "SDL_Plotter.h"
+
+#endif // CONSTANT_H_
