+++ Visual Studio/Tetris_VS/x64/Debug/Tetris_VS.log
@@ -1,2 +1,11 @@
-﻿  main.cpp
-  Tetris_VS.vcxproj -> C:\Users\reese\Documents\CSI1430-Tetris\Visual Studio\Tetris_VS\x64\Debug\Tetris_VS.exe
+﻿  SDL_Plotter.cpp
+c1xx : fatal  error C1083: Cannot open source file: '..\..\!Source Code\SDL Files\SDL_Plotter.cpp': No such file or directory
+  main.cpp
+C:\Users\reese\Documents\CSI1430-Tetris\!Source Code\main.cpp(25,5): error C2065: 'tile': undeclared identifier
+C:\Users\reese\Documents\CSI1430-Tetris\!Source Code\main.cpp(25,10): error C2146: syntax error: missing ';' before identifier 'square'
+C:\Users\reese\Documents\CSI1430-Tetris\!Source Code\main.cpp(25,10): error C2065: 'square': undeclared identifier
+C:\Users\reese\Documents\CSI1430-Tetris\!Source Code\main.cpp(32,35): error C2065: 'square': undeclared identifier
+C:\Users\reese\Documents\CSI1430-Tetris\!Source Code\main.cpp(34,34): error C2065: 'square': undeclared identifier
+C:\Users\reese\Documents\CSI1430-Tetris\!Source Code\main.cpp(38,9): error C2065: 'square': undeclared identifier
+C:\Users\reese\Documents\CSI1430-Tetris\!Source Code\main.cpp(39,9): error C2065: 'square': undeclared identifier
+  Generating Code...
